@@ -2,27 +2,28 @@ name: example
 description: A new Flutter application.
 version: 1.0.0+1
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 dependencies:
-  cloud_functions:
+  cloud_functions: '>=0.9.1-0'
   cupertino_icons:
-  firebase_auth:
+  firebase_auth: '>=0.21.0-0'
   firebase_core:
   firestore_ref:
     path: ..
   flutter:
     sdk: flutter
-  flutter_state_notifier:
-  freezed_annotation:
-  provider:
-  recase:
+  flutter_hooks: '>=0.16.0-0'
+  freezed_annotation: '>=0.13.0-0'
+  hooks_riverpod: '>=0.13.0-0'
+  recase: '>=4.0.0-0'
+  rxdart:
   state_notifier:
   subscription_holder: any
 dev_dependencies:
   build_runner: any
   flutter_test:
     sdk: flutter
-  freezed: any
+  freezed: '>=0.13.0-0'
   json_serializable: any
   pedantic_mono: any
 flutter:
