@@ -1,5 +1,5 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies
 
 part of 'user.dart';
 
@@ -12,45 +12,53 @@ User _$UserFromJson(Map<String, dynamic> json) {
   return _User.fromJson(json);
 }
 
+/// @nodoc
 class _$UserTearOff {
   const _$UserTearOff();
 
-// ignore: unused_element
   _User call(
-      {@required int count,
-      @TimestampConverter() DateTime createdAt,
-      @TimestampConverter() DateTime updatedAt}) {
+      {required int count,
+      @TimestampConverter() DateTime? createdAt,
+      @TimestampConverter() DateTime? updatedAt}) {
     return _User(
       count: count,
       createdAt: createdAt,
       updatedAt: updatedAt,
     );
   }
+
+  User fromJson(Map<String, Object> json) {
+    return User.fromJson(json);
+  }
 }
 
-// ignore: unused_element
+/// @nodoc
 const $User = _$UserTearOff();
 
+/// @nodoc
 mixin _$User {
   int get count;
   @TimestampConverter()
-  DateTime get createdAt;
+  DateTime? get createdAt;
   @TimestampConverter()
-  DateTime get updatedAt;
+  DateTime? get updatedAt;
 
   Map<String, dynamic> toJson();
+  @JsonKey(ignore: true)
   $UserCopyWith<User> get copyWith;
 }
 
+/// @nodoc
 abstract class $UserCopyWith<$Res> {
   factory $UserCopyWith(User value, $Res Function(User) then) =
       _$UserCopyWithImpl<$Res>;
   $Res call(
       {int count,
-      @TimestampConverter() DateTime createdAt,
-      @TimestampConverter() DateTime updatedAt});
+      @TimestampConverter() DateTime? createdAt,
+      @TimestampConverter() DateTime? updatedAt});
 }
 
+/// @nodoc
 class _$UserCopyWithImpl<$Res> implements $UserCopyWith<$Res> {
   _$UserCopyWithImpl(this._value, this._then);
 
@@ -60,30 +68,32 @@ class _$UserCopyWithImpl<$Res> implements $UserCopyWith<$Res> {
 
   @override
   $Res call({
-    Object count = freezed,
-    Object createdAt = freezed,
-    Object updatedAt = freezed,
+    Object? count = freezed,
+    Object? createdAt = freezed,
+    Object? updatedAt = freezed,
   }) {
     return _then(_value.copyWith(
       count: count == freezed ? _value.count : count as int,
       createdAt:
-          createdAt == freezed ? _value.createdAt : createdAt as DateTime,
+          createdAt == freezed ? _value.createdAt : createdAt as DateTime?,
       updatedAt:
-          updatedAt == freezed ? _value.updatedAt : updatedAt as DateTime,
+          updatedAt == freezed ? _value.updatedAt : updatedAt as DateTime?,
     ));
   }
 }
 
+/// @nodoc
 abstract class _$UserCopyWith<$Res> implements $UserCopyWith<$Res> {
   factory _$UserCopyWith(_User value, $Res Function(_User) then) =
       __$UserCopyWithImpl<$Res>;
   @override
   $Res call(
       {int count,
-      @TimestampConverter() DateTime createdAt,
-      @TimestampConverter() DateTime updatedAt});
+      @TimestampConverter() DateTime? createdAt,
+      @TimestampConverter() DateTime? updatedAt});
 }
 
+/// @nodoc
 class __$UserCopyWithImpl<$Res> extends _$UserCopyWithImpl<$Res>
     implements _$UserCopyWith<$Res> {
   __$UserCopyWithImpl(_User _value, $Res Function(_User) _then)
@@ -94,27 +104,28 @@ class __$UserCopyWithImpl<$Res> extends _$UserCopyWithImpl<$Res>
 
   @override
   $Res call({
-    Object count = freezed,
-    Object createdAt = freezed,
-    Object updatedAt = freezed,
+    Object? count = freezed,
+    Object? createdAt = freezed,
+    Object? updatedAt = freezed,
   }) {
     return _then(_User(
       count: count == freezed ? _value.count : count as int,
       createdAt:
-          createdAt == freezed ? _value.createdAt : createdAt as DateTime,
+          createdAt == freezed ? _value.createdAt : createdAt as DateTime?,
       updatedAt:
-          updatedAt == freezed ? _value.updatedAt : updatedAt as DateTime,
+          updatedAt == freezed ? _value.updatedAt : updatedAt as DateTime?,
     ));
   }
 }
 
 @JsonSerializable()
+
+/// @nodoc
 class _$_User with DiagnosticableTreeMixin implements _User {
   const _$_User(
-      {@required this.count,
+      {required this.count,
       @TimestampConverter() this.createdAt,
-      @TimestampConverter() this.updatedAt})
-      : assert(count != null);
+      @TimestampConverter() this.updatedAt});
 
   factory _$_User.fromJson(Map<String, dynamic> json) =>
       _$_$_UserFromJson(json);
@@ -123,10 +134,10 @@ class _$_User with DiagnosticableTreeMixin implements _User {
   final int count;
   @override
   @TimestampConverter()
-  final DateTime createdAt;
+  final DateTime? createdAt;
   @override
   @TimestampConverter()
-  final DateTime updatedAt;
+  final DateTime? updatedAt;
 
   @override
   String toString({DiagnosticLevel minLevel = DiagnosticLevel.info}) {
@@ -164,6 +175,7 @@ class _$_User with DiagnosticableTreeMixin implements _User {
       const DeepCollectionEquality().hash(createdAt) ^
       const DeepCollectionEquality().hash(updatedAt);
 
+  @JsonKey(ignore: true)
   @override
   _$UserCopyWith<_User> get copyWith =>
       __$UserCopyWithImpl<_User>(this, _$identity);
@@ -176,9 +188,9 @@ class _$_User with DiagnosticableTreeMixin implements _User {
 
 abstract class _User implements User {
   const factory _User(
-      {@required int count,
-      @TimestampConverter() DateTime createdAt,
-      @TimestampConverter() DateTime updatedAt}) = _$_User;
+      {required int count,
+      @TimestampConverter() DateTime? createdAt,
+      @TimestampConverter() DateTime? updatedAt}) = _$_User;
 
   factory _User.fromJson(Map<String, dynamic> json) = _$_User.fromJson;
 
@@ -186,10 +198,11 @@ abstract class _User implements User {
   int get count;
   @override
   @TimestampConverter()
-  DateTime get createdAt;
+  DateTime? get createdAt;
   @override
   @TimestampConverter()
-  DateTime get updatedAt;
+  DateTime? get updatedAt;
   @override
+  @JsonKey(ignore: true)
   _$UserCopyWith<_User> get copyWith;
 }
