@@ -2,15 +2,16 @@ import 'package:example/app.dart';
 import 'package:example/router.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
+import 'package:flutter_hooks/flutter_hooks.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-class HomePage extends StatelessWidget {
-  const HomePage({Key key}) : super(key: key);
+class HomePage extends HookWidget {
+  const HomePage({Key? key}) : super(key: key);
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text(context.select((AppInfo info) => info.title)),
+        title: Text(useProvider(appInfo).title),
       ),
       body: ListView(
         children: PageInfo.all.map((info) {
