@@ -2,6 +2,7 @@ import 'package:example/pages/home_page.dart';
 import 'package:example/pages/paging_page/paging_page.dart';
 import 'package:example/pages/user_counter_page.dart';
 import 'package:flutter/material.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:recase/recase.dart';
 
 import 'util/util.dart';
@@ -11,23 +12,25 @@ typedef WidgetPageBuilder = Widget Function(
   RouteSettings settings,
 );
 
+final router = Provider((_) => Router());
+
 // ignore: avoid_classes_with_only_static_members
 class Router {
   static const root = '/';
 
   final _routes = <String, WidgetPageBuilder>{
     root: (_, __) => const HomePage(),
-    UserCounterPage.routeName: (_, __) => UserCounterPage.wrapped(),
-    PagingPage.routeName: (_, __) => PagingPage.wrapped(),
+    UserCounterPage.routeName: (_, __) => const UserCounterPage(),
+    PagingPage.routeName: (_, __) => const PagingPage(),
   };
   final _modalRoutes = <String, WidgetPageBuilder>{};
 
-  Route<dynamic> onGenerateRoute(RouteSettings settings) {
+  Route<dynamic>? onGenerateRoute(RouteSettings settings) {
     logger.info(settings.name);
     var pageBuilder = _routes[settings.name];
     if (pageBuilder != null) {
       return MaterialPageRoute<void>(
-        builder: (context) => pageBuilder(context, settings),
+        builder: (context) => pageBuilder!(context, settings),
         settings: settings,
       );
     }
@@ -35,7 +38,7 @@ class Router {
     pageBuilder = _modalRoutes[settings.name];
     if (pageBuilder != null) {
       return MaterialPageRoute<void>(
-        builder: (context) => pageBuilder(context, settings),
+        builder: (context) => pageBuilder!(context, settings),
         settings: settings,
         fullscreenDialog: true,
       );
@@ -51,7 +54,7 @@ String pascalCaseFromRouteName(String name) => name.substring(1).pascalCase;
 @immutable
 class PageInfo {
   const PageInfo({
-    @required this.routeName,
+    required this.routeName,
   });
 
   final String routeName;
