@@ -1,3 +1,3 @@
 export 'authenticator.dart';
-export 'user_notifier.dart';
+export 'my_notifier.dart';
 export 'users_notofier.dart';
